+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,7 +1,6 @@
 package org.apache.catalina;
 
 import java.io.IOException;
-import nextstep.jwp.model.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
