+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestLine.java
@@ -1,19 +1,19 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
-public class RequestLine {
+public class HttpRequestLine {
 
     private final String method;
     private final String target;
     private final String version;
 
-    private RequestLine(final String method, final String target, final String version) {
+    private HttpRequestLine(final String method, final String target, final String version) {
         this.method = method;
         this.target = target;
         this.version = version;
     }
 
-    public static RequestLine from(final String method, final String target, final String version) {
-        return new RequestLine(method, target, version);
+    public static HttpRequestLine from(final String method, final String target, final String version) {
+        return new HttpRequestLine(method, target, version);
     }
 
     public String getMethod() {
