+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponse.java
@@ -1,14 +1,16 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
+
+import org.apache.coyote.http11.HttpHeaderType;
+import org.apache.coyote.http11.HttpHeaders;
 
 import java.util.Map;
 
-import static org.apache.coyote.http11.HttpStatusCode.NOT_FOUND;
+import static org.apache.coyote.http11.response.HttpStatusCode.NOT_FOUND;
 
 public class HttpResponse {
 
     private HttpStatusCode statusCode;
     private final HttpHeaders headers;
-//    private final Map<String, String> header;
     private String body;
 
     private HttpResponse(final HttpStatusCode statusCode, final HttpHeaders headers, final String body) {
